issue #95: asynchron output for servlet 3.1
The handling of a play result in the default request handler is moved
to pushPlayResultToServletOS. This handling stayed the same. This change
should not have any impact on the existing servlet wrapper for servlet < 3.1.

For servlet 3.1, pushPlayResultToServletOS is overwritten to provide a
non-blocking asynchron handling based on the new WriteListener.